replace script to another folder
