+++ Assets/Sandbox/Amid/Scenes/lvl2.unity
@@ -14,7 +14,7 @@ OcclusionCullingSettings:
 RenderSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 8
-  m_Fog: 1
+  m_Fog: 0
   m_FogColor: {r: 0.8897059, g: 0.8897059, b: 0.8897059, a: 1}
   m_FogMode: 3
   m_FogDensity: 0.015
@@ -2484,7 +2484,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 50750481671431f48b35f1d65ace61de, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  singleTrigger: 1
+  _singleTrigger: 1
 --- !u!82 &435333995
 AudioSource:
   m_ObjectHideFlags: 0
